python/fastApi/pyproject.toml
[tool.poetry]
name = "crud"
version = "0.1.0"
description = ""
authors = ["Saul Burgos <saulburgos6@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.99.1"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
